Update
1. Changed SpriteLoader to experiment with doubles for more precision.  No difference.  Will probably revert.  There's no escaping pixel perfect ortho sizes.
2. Changed camera speed, default ortho settings, and added a function to zoom with world scale.  Also won't work.
3. Tweaked the box select interface to behave the way I actually want it to.  Also merged both types of select into a single function with a bool passed in to determine which type of select is required.

+++ Tile Map Testing/Assets/Scripts/MonoBehavior/SpriteLoader.cs
@@ -37,8 +37,8 @@ public class SpriteLoader : MonoBehaviour {
 	}
 
 	void InitializeSpriteUVs(){
-		float spriteWidth = 0f;
-		float spriteHeight = 0f;
+		double spriteWidth = 0d;
+		double spriteHeight = 0d;
 
 		//Calculate spritesheet width
 		foreach (Sprite sprite in SpriteArray) {
@@ -53,14 +53,14 @@ public class SpriteLoader : MonoBehaviour {
 		foreach (Sprite sprite in SpriteArray) {
 			Vector2[] uvs = new Vector2[4];
 
-			uvs [0] = new Vector2 (sprite.rect.x / spriteWidth, sprite.rect.y / spriteHeight);
-			uvs [1] = new Vector2 ((sprite.rect.x + sprite.rect.width) / spriteWidth, sprite.rect.y / spriteHeight);
-			uvs [2] = new Vector2 (sprite.rect.x / spriteWidth, (sprite.rect.y + sprite.rect.height) / spriteHeight);
-			uvs [3] = new Vector2 ((sprite.rect.x + sprite.rect.width) / spriteWidth, (sprite.rect.y + sprite.rect.height) / spriteHeight);
+			uvs [0] = new Vector2 ((float)(sprite.rect.x / spriteWidth), (float)(sprite.rect.y / spriteHeight));
+			uvs [1] = new Vector2 ((float)((sprite.rect.x + sprite.rect.width) / spriteWidth), (float)(sprite.rect.y / spriteHeight));
+			uvs [2] = new Vector2 ((float)(sprite.rect.x / spriteWidth), (float)((sprite.rect.y + sprite.rect.height) / spriteHeight));
+			uvs [3] = new Vector2 ((float)((sprite.rect.x + sprite.rect.width) / spriteWidth), (float)((sprite.rect.y + sprite.rect.height) / spriteHeight));
 
 			uvMap.Add (sprite.name, uvs);
 		}
-	}		
+	}
 
 	public Vector2[] GetWorldUVS (Tile tile){
 		string key = tile.FLOOR.ToString ();
